src/requirements.txt
apache-airflow[crypto,postgres]==1.10.5
pika==0.12.0
colour==0.1.5
flake8==3.5.0
pytest==3.7.4
pytest-cov==2.5.1
